Move some model pipelines to pipelines package
