+++ magenta/models/music_vae/data_hierarchical.py
@@ -21,7 +21,7 @@ from __future__ import print_function
 import abc
 
 from magenta.models.music_vae import data
-from magenta.models.performance_rnn import performance_rnn_pipeline
+from magenta.pipelines import performance_rnn_pipeline
 import magenta.music as mm
 from magenta.music import chords_lib
 from magenta.music import performance_lib
