+++ magenta/models/performance_rnn/performance_rnn_create_dataset.py
@@ -23,8 +23,8 @@ NoteSequence within a limited range.
 import os
 
 from magenta.models.performance_rnn import performance_model
-from magenta.models.performance_rnn import performance_rnn_pipeline
 from magenta.pipelines import pipeline
+from magenta.pipelines import performance_rnn_pipeline
 import tensorflow as tf
 
 flags = tf.app.flags
