+++ magenta/models/pianoroll_rnn_nade/pianoroll_rnn_nade_create_dataset_test.py
@@ -15,9 +15,9 @@
 """Tests for pianoroll_rnn_nade_create_dataset."""
 
 import magenta
-from magenta.models.pianoroll_rnn_nade import pianoroll_rnn_nade_pipeline
 from magenta.models.shared import events_rnn_model
 import magenta.music as mm
+from magenta.pipelines import pianoroll_rnn_nade_pipeline
 from magenta.protobuf import music_pb2
 import tensorflow as tf
