+++ magenta/pipelines/performance_rnn_pipeline.py
