+++ magenta/pipelines/pianoroll_rnn_nade_pipeline.py
