+++ magenta/pipelines/pianoroll_rnn_nade_pipeline_test.py
@@ -14,9 +14,9 @@
 
 """Tests for pianoroll_lib."""
 
-from magenta.models.pianoroll_rnn_nade import pianoroll_rnn_nade_pipeline
 from magenta.music import sequences_lib
 from magenta.music import testing_lib as music_testing_lib
+from magenta.pipelines import pianoroll_rnn_nade_pipeline
 from magenta.protobuf import music_pb2
 import tensorflow as tf
